imported java.util.Queue package and restructured the methods
