+++ java/ContainerWithMostWater.java
@@ -1,0 +1,24 @@
+public class ContainerWithMostWater {
+    public int maxArea(int[] height) {
+        int left = 0, right = height.length - 1;
+        int maxArea = 0;
+
+        while (left < right) {
+            int area = Math.min(height[left], height[right]) * (right - left);
+            maxArea = Math.max(maxArea, area);
+
+            if (height[left] < height[right]) {
+                left++;
+            } else {
+                right--;
+            }
+        }
+
+        return maxArea;
+    }
+
+    public static void main(String[] args) {
+        ContainerWithMostWater obj = new ContainerWithMostWater();
+        System.out.println(obj.maxArea(new int[]{1,8,6,2,5,4,8,3,7})); // 49
+    }
+}
